+++ lab9_project/vec2d.h
@@ -1,17 +1,19 @@
 #ifndef VEC2D
 #define VEC2D
 
+#include <stdint.h>
+
 // Vectors
 typedef struct {
-    int x;
-    int y;
+  uint8_t x;
+  uint8_t y;
 } vec2d;
 
-static vec2d makeVec(int x, int y){
-    vec2d vector;
-    vector.x = x;
-    vector.y = y;
-    return vector;
+static vec2d makeVec(uint8_t x, uint8_t y) {
+  vec2d vector;
+  vector.x = x;
+  vector.y = y;
+  return vector;
 }
 
 #endif
